Change TransferAndCall, convert amount to 18 decimals before calling allow tokens
Transfer and call now calls using ERC777 and then calls using ERC6777 standar, tokens that implement both will have to controll this. Event is emited following both ERC rules

When verifying the amount we now need to convert it to 18 decimals as its the value used for the comparison
Test added and modifyed accordingly.